fix(api): Introduce CommonInstancetypesDeployment
This is a standalone configurable used to control the overall deployment
of the common-instancetypes cluster-wide resources by KubeVirt. This
sits outside of the specific instancetypeConfiguration configurable
within the KubeVirt CR that controls other core functionality related to
instance types and preferences in KubeVirt and that separation is
mirrored here in HCO with separate configurables.

While this introduces a new configurable to the API it is being classed
as a bugfix as without this there is no supported way for users to
disable the deployment of common-instancetypes within HCO deployed
environments.

NOTE(lyarwood): This backport to 1.13 reworks the change as the
associated CommonInstancetypesDeployment struct is not available for us
to use in KubeVirt 1.3.0. As such this change inlines the struct within
the HCO schema and uses the KubeVirt FG to control deployment of
common-instancetypes. This should be backwardly compatible with HCO >= 1.14.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
(cherry picked from commit ba487f9746e4ec6dd9f7faacbda485af5be31bee)
(cherry picked from commit 0844710958bbfcba788f7638fa3a561919008623)